+++ packages/init/src/webframeworks/const.ts
@@ -1,0 +1,10 @@
+import { PACKAGE_VERSION } from "../lib.ts";
+
+export const defaultDevDependencies = {
+  "@fedify/lint": PACKAGE_VERSION,
+  "eslint": "^9.0.0",
+};
+
+export const defaultDenoDependencies = {
+  "@fedify/lint": PACKAGE_VERSION,
+};